Moved some code into module aflutter
